requirements_dev.txt
-r requirements.txt

pytest==6.2.1
pytest-django==4.1.0
mixer==7.0.6
django-debug-toolbar==2.2
flake8==3.8.3
isort==5.7.0
